chore: have types in one file
