+++ jest.config.ts
@@ -1,5 +1,5 @@
 export default {
-	transform: {
-		'^.+\\.(t|j)sx?$': ['@swc/jest'],
-	},
-}
+  transform: {
+    "^.+\\.(t|j)sx?$": ["@swc/jest"],
+  },
+};
