+++ package.json
@@ -1,63 +1,63 @@
 {
-	"name": "hexagon-generator",
-	"version": "1.0.0",
-	"description": "",
-	"source": "html/index.html",
-	"scripts": {
-		"format:fix": "prettier --config .prettierrc 'ts/**/*.ts' --write",
-		"lint": "eslint . --ext .ts",
-		"lint:fix": "eslint . --ext .ts --fix",
-		"prepare": "husky install",
-		"build:dev": "webpack serve --open --config webpack.dev.js",
-		"build:prod": "webpack --config webpack.prod.js",
-		"test": "jest",
-		"test:coverage": "jest --coverage",
-		"test:all": "jest --watchAll --verbose"
-	},
-	"keywords": [
-		"typescript",
-		"html5",
-		"css3",
-		"webpack"
-	],
-	"author": "WV",
-	"license": "ISC",
-	"devDependencies": {
-		"@swc/core": "^1.3.44",
-		"@swc/jest": "^0.2.29",
-		"@types/jest": "^27.4.1",
-		"@types/node": "^17.0.45",
-		"@types/webpack": "^5.28.1",
-		"@typescript-eslint/eslint-plugin": "^5.59.2",
-		"@typescript-eslint/parser": "^5.60.1",
-		"@webpack-cli/generators": "^2.5.0",
-		"clean-webpack-plugin": "^4.0.0",
-		"copy-webpack-plugin": "^10.2.4",
-		"css-loader": "^6.7.1",
-		"declaration-bundler-webpack-plugin": "^1.0.3",
-		"eslint": "^8.44.0",
-		"eslint-config-prettier": "^8.10.0",
-		"eslint-plugin-no-loops": "^0.3.0",
-		"eslint-plugin-prettier": "^4.2.1",
-		"git-commit-msg-linter": "^4.11.1",
-		"html-webpack-plugin": "^5.5.0",
-		"husky": "^7.0.4",
-		"jest": "^27.5.1",
-		"lint-staged": "^12.5.0",
-		"prettier": "^2.8.8",
-		"style-loader": "^3.3.1",
-		"ts-jest": "^27.1.5",
-		"ts-loader": "^9.4.4",
-		"typescript": "^4.9.5",
-		"webpack": "^5.88.2",
-		"webpack-cli": "^4.10.0",
-		"webpack-dev-server": "^4.8.1",
-		"webpack-merge": "^5.8.0"
-	},
-	"dependencies": {
-		"ts-node": "^10.9.1"
-	},
-	"lint-staged": {
-		"*.ts": "eslint --cache --fix"
-	}
+  "name": "hexagon-generator",
+  "version": "1.0.0",
+  "description": "",
+  "source": "html/index.html",
+  "scripts": {
+    "format:fix": "prettier --config .prettierrc 'ts/**/*.ts' --write",
+    "lint": "eslint . --ext .ts",
+    "lint:fix": "eslint . --ext .ts --fix",
+    "prepare": "husky install",
+    "build:dev": "webpack serve --open --config webpack.dev.js",
+    "build:prod": "webpack --config webpack.prod.js",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
+    "test:all": "jest --watchAll --verbose"
+  },
+  "keywords": [
+    "typescript",
+    "html5",
+    "css3",
+    "webpack"
+  ],
+  "author": "WV",
+  "license": "ISC",
+  "devDependencies": {
+    "@swc/core": "^1.3.44",
+    "@swc/jest": "^0.2.29",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.45",
+    "@types/webpack": "^5.28.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.60.1",
+    "@webpack-cli/generators": "^2.5.0",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "css-loader": "^6.7.1",
+    "declaration-bundler-webpack-plugin": "^1.0.3",
+    "eslint": "^8.44.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-no-loops": "^0.3.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "git-commit-msg-linter": "^4.11.1",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^7.0.4",
+    "jest": "^27.5.1",
+    "lint-staged": "^12.5.0",
+    "prettier": "^2.8.8",
+    "style-loader": "^3.3.1",
+    "ts-jest": "^27.1.5",
+    "ts-loader": "^9.4.4",
+    "typescript": "^4.9.5",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.8.1",
+    "webpack-merge": "^5.8.0"
+  },
+  "dependencies": {
+    "ts-node": "^10.9.1"
+  },
+  "lint-staged": {
+    "*.ts": "eslint --cache --fix"
+  }
 }
