+++ ts/logic/inputs/factory.ts
@@ -1,10 +1,10 @@
+import { Input, InputKind } from "../types";
 import { InputBase } from "./InputBase";
 import { InputFirstRow } from "./InputFirstRow";
 import { InputGap } from "./InputGap";
 import { InputMediaQuery } from "./InputMediaQuery";
 import { InputNumberOfHexagons } from "./InputNumberOfHexagons";
 import { InputRoot } from "./InputRoot";
-import { Input, InputKind } from "./types";
 
 export class InputFactory {
   createInput(
