+++ ts/logic/inputs/inputElements.ts
@@ -1,10 +1,10 @@
+import { InputKind } from "../types";
 import { InputFactory } from "./factory";
 import { InputFirstRow } from "./InputFirstRow";
 import { InputGap } from "./InputGap";
 import { InputMediaQuery } from "./InputMediaQuery";
 import { InputNumberOfHexagons } from "./InputNumberOfHexagons";
 import { InputRoot } from "./InputRoot";
-import { InputKind } from "./types";
 
 const factory = new InputFactory();
